Move theme to proper file
