+++ client/src/constants/styleTheme.js
@@ -1,0 +1,521 @@
+import { FontFamily } from "./styleConstants";
+
+export const MaterialThemeOceanic = {
+    name: "MaterialThemeOceanic",
+    FontFamily,
+
+    Background: "#263238",
+    Foreground: "#B0BEC5",
+    Text: "#607D8B",
+    SelectionBackground: "#546E7A",
+    SelectionForeground: "#FFFFFF",
+    Buttons: "#2E3C43",
+    SecondBackground: "#32424A",
+    Disabled: "#415967",
+    Contrast: "#1E272C",
+    Active: "#314549",
+    Border: "#2A373E",
+    Highlight: "#425B67",
+    Tree: "#546E7A",
+    Notifications: "#1E272C",
+    AccentColor: "#009688",
+    ExcludedFilesColor: "#2E3C43",
+    CommentsColor: "#546e7a",
+    LinksColor: "#80cbc4",
+    FunctionsColor: "#82aaff",
+    KeywordsColor: "#c792ea",
+    TagsColor: "#f07178",
+    StringsColor: "#c3e88d",
+    OperatorsColor: "#89ddff",
+    AttributesColor: "#ffcb6b",
+    NumbersColor: "#f78c6c",
+    ParametersColor: "#f78c6c",
+
+    // Custom Color
+    SecondText: "#FFFFFF",
+    ContrastBackground: "#000000",
+};
+
+export const MaterialThemePalenight = {
+    name: "MaterialThemePalenight",
+    FontFamily,
+
+    Background: "#292D3E",
+    Foreground: "#A6ACCD",
+    Text: "#676E95",
+    SelectionBackground: "#3C435E",
+    SelectionForeground: "#FFFFFF",
+    Buttons: "#303348",
+    SecondBackground: "#34324a",
+    Disabled: "#515772",
+    Contrast: "#202331",
+    Active: "#414863",
+    Border: "#2b2a3e",
+    Highlight: "#444267",
+    Tree: "#676E95",
+    Notifications: "#202331",
+    AccentColor: "#ab47bc",
+    ExcludedFilesColor: "#2f2e43",
+    CommentsColor: "#676E95",
+    LinksColor: "#80cbc4",
+    FunctionsColor: "#82aaff",
+    KeywordsColor: "#c792ea",
+    TagsColor: "#f07178",
+    StringsColor: "#c3e88d",
+    OperatorsColor: "#89ddff",
+    AttributesColor: "#ffcb6b",
+    NumbersColor: "#f78c6c",
+    ParametersColor: "#f78c6c",
+
+    // Custom Color
+    SecondText: "#FFFFFF",
+    ContrastBackground: "#000000",
+};
+
+export const MaterialThemeGithub = {
+    name: "MaterialThemeGithub",
+    FontFamily,
+
+    Background: "#F7F8FA",
+    Foreground: "#5B6168",
+    Text: "#292D31",
+    SelectionBackground: "#3A66D1",
+    SelectionForeground: "#FFFFFF",
+    Buttons: "#F1F3F6",
+    SecondBackground: "#F7F8FA",
+    Disabled: "#6C737C",
+    Contrast: "#FFFFFF",
+    Active: "#cce5ff",
+    Border: "#DFE1E4",
+    Highlight: "#CCE5FF",
+    Tree: "#3A66D1",
+    Notifications: "#DFECFE",
+    AccentColor: "#79CB60",
+    ExcludedFilesColor: "#FDFDFD",
+    CommentsColor: "#6A737D",
+    LinksColor: "#005CC5",
+    FunctionsColor: "#6F42C1",
+    KeywordsColor: "#D73A49",
+    TagsColor: "#22863A",
+    StringsColor: "#032F62",
+    OperatorsColor: "#D73A49",
+    AttributesColor: "#6F42C1",
+    NumbersColor: "#005CC5",
+    ParametersColor: "#E36209",
+
+    // Custom Color
+    SecondText: "#000000",
+    ContrastBackground: "#FFFFFF",
+};
+
+export const MaterialThemeDarker = {
+    name: "MaterialThemeDarker",
+    FontFamily,
+
+    Background: "#212121",
+    Foreground: "#B0BEC5",
+    Text: "#727272",
+    SelectionBackground: "#353535",
+    SelectionForeground: "#FFFFFF",
+    Buttons: "#2A2A2A",
+    SecondBackground: "#292929",
+    Disabled: "#474747",
+    Contrast: "#1A1A1A",
+    Active: "#323232",
+    Border: "#292929",
+    Highlight: "#3F3F3F",
+    Tree: "#323232",
+    Notifications: "#1A1A1A",
+    AccentColor: "#FF9800",
+    ExcludedFilesColor: "#323232",
+    CommentsColor: "#616161",
+    LinksColor: "#80cbc4",
+    FunctionsColor: "#82aaff",
+    KeywordsColor: "#c792ea",
+    TagsColor: "#f07178",
+    StringsColor: "#c3e88d",
+    OperatorsColor: "#89ddff",
+    AttributesColor: "#ffcb6b",
+    NumbersColor: "#f78c6c",
+    ParametersColor: "#f78c6c",
+
+    // Custom Color
+    SecondText: "#FFFFFF",
+    ContrastBackground: "#000000",
+};
+
+export const MaterialThemeDeepOcean = {
+    name: "MaterialThemeDeepOcean",
+    FontFamily,
+
+    Background: "#0F111A",
+    Foreground: "#8F93A2",
+    Text: "#4B526D",
+    SelectionBackground: "#232632",
+    SelectionForeground: "#FFFFFF",
+    Buttons: "#191A21",
+    SecondBackground: "#181A1F",
+    Disabled: "#464B5D",
+    Contrast: "#090B10",
+    Active: "#1A1C25",
+    Border: "#0F111A",
+    Highlight: "#1F2233",
+    Tree: "#717CB4",
+    Notifications: "#090B10",
+    AccentColor: "#84ffff",
+    ExcludedFilesColor: "#292D3E",
+    CommentsColor: "#717CB4",
+    LinksColor: "#80cbc4",
+    FunctionsColor: "#82aaff",
+    KeywordsColor: "#c792ea",
+    TagsColor: "#f07178",
+    StringsColor: "#c3e88d",
+    OperatorsColor: "#89ddff",
+    AttributesColor: "#ffcb6b",
+    NumbersColor: "#f78c6c",
+    ParametersColor: "#F59762",
+
+    // Custom Color
+    SecondText: "#FFFFFF",
+    ContrastBackground: "#000000",
+};
+
+export const ThemeMonoKaiPro = {
+    name: "ThemeMonoKaiPro",
+    FontFamily,
+
+    Background: "#2D2A2E",
+    Foreground: "#fcfcfa",
+    Text: "#939293",
+    SelectionBackground: "#5B595C",
+    SelectionForeground: "#ffd866",
+    Buttons: "#403E41",
+    SecondBackground: "#403E41",
+    Disabled: "#727072",
+    Contrast: "#221F22",
+    Active: "#4A474B",
+    Border: "#2d2a2e",
+    Highlight: "#5b595c",
+    Tree: "#403E41",
+    Notifications: "#363437",
+    AccentColor: "#ffd866",
+    ExcludedFilesColor: "#3a3a3c",
+    CommentsColor: "#727072",
+    LinksColor: "#78DCE8",
+    FunctionsColor: "#A9DC76",
+    KeywordsColor: "#FF6188",
+    TagsColor: "#FF6188",
+    StringsColor: "#FFD866",
+    OperatorsColor: "#FF6188",
+    AttributesColor: "#78DCE8",
+    NumbersColor: "#AB9DF2",
+    ParametersColor: "#F59762",
+
+    // Custom Color
+    SecondText: "#FFFFFF",
+    ContrastBackground: "#000000",
+};
+
+export const MaterialThemeLighter = {
+    name: "MaterialThemeLighter",
+    FontFamily,
+
+    Background: "#FAFAFA",
+    Foreground: "#546E7A",
+    Text: "#94A7B0",
+    SelectionBackground: "#80CBC4",
+    SelectionForeground: "#546e7a",
+    Buttons: "#F3F4F5",
+    SecondBackground: "#eae8e8",
+    Disabled: "#D2D4D5",
+    Contrast: "#F4F4F4",
+    Active: "#E7E7E8",
+    Border: "#d3e1e8",
+    Highlight: "#E7E7E8",
+    Tree: "#80CBC4",
+    Notifications: "#eae8e8",
+    AccentColor: "#00BCD4",
+    ExcludedFilesColor: "#eae8e8",
+    CommentsColor: "#AABFC9",
+    LinksColor: "#39ADB5",
+    FunctionsColor: "#6182B8",
+    KeywordsColor: "#7C4DFF",
+    TagsColor: "#E53935",
+    StringsColor: "#91B859",
+    OperatorsColor: "#39ADB5",
+    AttributesColor: "#F6A434",
+    NumbersColor: "#F76D47",
+    ParametersColor: "#F76D47",
+
+    // Custom Color
+    SecondText: "#000000",
+    ContrastBackground: "#FFFFFF",
+};
+
+export const ThemeOneLight = {
+    name: "ThemeOneLight",
+    FontFamily,
+
+    Background: "#FAFAFA",
+    Foreground: "#232324",
+    Text: "#9D9D9F",
+    SelectionBackground: "#DBDBDC",
+    SelectionForeground: "#232324",
+    Buttons: "#DBDBDC",
+    SecondBackground: "#EAEAEB",
+    Disabled: "#424243",
+    Contrast: "#FFFFFF",
+    Active: "#DBDBDC",
+    Border: "#DBDBDC",
+    Highlight: "#EAEAEB",
+    Tree: "#DBDBDC",
+    Notifications: "#F2F2F2",
+    AccentColor: "#2979ff",
+    ExcludedFilesColor: "#DBDBDC",
+    CommentsColor: "#A0A1A7",
+    LinksColor: "#0184BC",
+    FunctionsColor: "#4078F2",
+    KeywordsColor: "#A626A4",
+    TagsColor: "#E4564A",
+    StringsColor: "#50A14E",
+    OperatorsColor: "#4078F2",
+    AttributesColor: "#C18401",
+    NumbersColor: "#986801",
+    ParametersColor: "#383A42",
+
+    // Custom Color
+    SecondText: "#000000",
+    ContrastBackground: "#FFFFFF",
+};
+
+export const SolarizedLight = {
+    name: "SolarizedLight",
+    FontFamily,
+
+    Background: "#fdf6e3",
+    Foreground: "#586e75",
+    Text: "#93a1a1",
+    SelectionBackground: "#93a1a1",
+    SelectionForeground: "#002b36",
+    Buttons: "#FEFBF1",
+    SecondBackground: "#F6F0DE",
+    Disabled: "#C9CCC3",
+    Contrast: "#eee8d5",
+    Active: "#F6F0DE",
+    Border: "#edead9",
+    Highlight: "#F6F0DE",
+    Tree: "#93a1a1",
+    Notifications: "#EDE8D4",
+    AccentColor: "#d33682",
+    ExcludedFilesColor: "#E3DCC9",
+    CommentsColor: "#93A1A1",
+    LinksColor: "#268BD2",
+    FunctionsColor: "#B58900",
+    KeywordsColor: "#859900",
+    TagsColor: "#268BD2",
+    StringsColor: "#2AA198",
+    OperatorsColor: "#657B83",
+    AttributesColor: "#657B83",
+    NumbersColor: "#D33682",
+    ParametersColor: "#657B83",
+
+    // Custom Color
+    SecondText: "#000000",
+    ContrastBackground: "#FFFFFF",
+};
+
+export const ThemeDracula = {
+    name: "ThemeDracula",
+    FontFamily,
+
+    Background: "#282A36",
+    Foreground: "#F8F8F2",
+    Text: "#6272A4",
+    SelectionBackground: "#44475A",
+    SelectionForeground: "#8BE9FD",
+    Buttons: "#393C4B",
+    SecondBackground: "#282A36",
+    Disabled: "#6272A4",
+    Contrast: "#191A21",
+    Active: "#44475A",
+    Border: "#21222C",
+    Highlight: "#6272A4",
+    Tree: "#44475A",
+    Notifications: "#1D2228",
+    AccentColor: "#FF79C5",
+    ExcludedFilesColor: "#34353D",
+    CommentsColor: "#6272A4",
+    LinksColor: "#F1FA8C",
+    FunctionsColor: "#50FA78",
+    KeywordsColor: "#FF79C6",
+    TagsColor: "#FF79C6",
+    StringsColor: "#F1FA8C",
+    OperatorsColor: "#FF79C6",
+    AttributesColor: "#50FA7B",
+    NumbersColor: "#BD93F9",
+    ParametersColor: "#FFB86C",
+
+    // Custom Color
+    SecondText: "#FFFFFF",
+    ContrastBackground: "#000000",
+};
+
+export const ThemeArcDark = {
+    name: "ThemeArcDark",
+    FontFamily,
+
+    Background: "#2f343f",
+    Foreground: "#D3DAE3",
+    Text: "#8b9eb5",
+    SelectionBackground: "#8888FF",
+    SelectionForeground: "#FFFFFF",
+    Buttons: "#383C4A",
+    SecondBackground: "#393f4c",
+    Disabled: "#D3DAE3",
+    Contrast: "#262b33",
+    Active: "#41416A",
+    Border: "#404552",
+    Highlight: "#3F3F46",
+    Tree: "#08507C",
+    Notifications: "#262a33",
+    AccentColor: "#42A5F5",
+    ExcludedFilesColor: "#474B57",
+    CommentsColor: "#747C84",
+    LinksColor: "#7587A6",
+    FunctionsColor: "#7587A6",
+    KeywordsColor: "#9B859D",
+    TagsColor: "#CF6A4C",
+    StringsColor: "#8F9D6A",
+    OperatorsColor: "#A7A7A7",
+    AttributesColor: "#F9EE98",
+    NumbersColor: "#CDA869",
+    ParametersColor: "#CDA869",
+
+    // Custom Color
+    SecondText: "#FFFFFF",
+    ContrastBackground: "#000000",
+};
+
+export const ThemeOneDark = {
+    name: "ThemeOneDark",
+    FontFamily,
+
+    Background: "#282C34",
+    Foreground: "#979FAD",
+    Text: "#979FAD",
+    SelectionBackground: "#3A3F4B",
+    SelectionForeground: "#FFFFFF",
+    Buttons: "#3A3F4B",
+    SecondBackground: "#2F333D",
+    Disabled: "#6B727D",
+    Contrast: "#21252B",
+    Active: "#383E49",
+    Border: "#282C34",
+    Highlight: "#383D48",
+    Tree: "#3A3F4B",
+    Notifications: "#282C34",
+    AccentColor: "#2979ff",
+    ExcludedFilesColor: "#3c4150",
+    CommentsColor: "#59626F",
+    LinksColor: "#56B6C2",
+    FunctionsColor: "#61AEEF",
+    KeywordsColor: "#C679DD",
+    TagsColor: "#F07178",
+    StringsColor: "#98C379",
+    OperatorsColor: "#61AFEF",
+    AttributesColor: "#E5C17C",
+    NumbersColor: "#D19A66",
+    ParametersColor: "#ABB2BF",
+
+    // Custom Color
+    SecondText: "#FFFFFF",
+    ContrastBackground: "#000000",
+};
+
+export const ThemeNightOwl = {
+    name: "ThemeNightOwl",
+    FontFamily,
+
+    Background: "#011627",
+    Foreground: "#d6deeb",
+    Text: "#5f7e97",
+    SelectionBackground: "#1d3b53FF",
+    SelectionForeground: "#FFFFFF",
+    Buttons: "#0b253a",
+    SecondBackground: "#0b2942",
+    Disabled: "#697098",
+    Contrast: "#010e1a",
+    Active: "#234d70",
+    Border: "#122d42",
+    Highlight: "#084d81",
+    Tree: "#234d70",
+    Notifications: "#01111d",
+    AccentColor: "#7e57c2",
+    ExcludedFilesColor: "#0e293f",
+    CommentsColor: "#637777",
+    LinksColor: "#ecc48d",
+    FunctionsColor: "#B58900",
+    KeywordsColor: "#c792ea",
+    TagsColor: "#7fdbca",
+    StringsColor: "#ecc48d",
+    OperatorsColor: "#c792ea",
+    AttributesColor: "#ef5350",
+    NumbersColor: "#f78c6c",
+    ParametersColor: "#d6deeb",
+
+    // Custom Color
+    SecondText: "#FFFFFF",
+    ContrastBackground: "#000000",
+};
+
+export const ThemeLightOwl = {
+    name: "ThemeLightOwl",
+    FontFamily,
+
+    Background: "#F0F0F0",
+    Foreground: "#403f53",
+    Text: "#90A7B2",
+    SelectionBackground: "#E0E0E0",
+    SelectionForeground: "#403f53",
+    Buttons: "#d9d9d9",
+    SecondBackground: "#FBFBFB",
+    Disabled: "#93A1A1",
+    Contrast: "#D0D0D0",
+    Active: "#d3e8f8",
+    Border: "#d9d9d9",
+    Highlight: "#CCCCCC",
+    Tree: "#7cd3e8",
+    Notifications: "#F0F0F0",
+    AccentColor: "#2AA298",
+    ExcludedFilesColor: "#E0E7EA",
+    CommentsColor: "#989fb1",
+    LinksColor: "#ff869a",
+    FunctionsColor: "#4876d6",
+    KeywordsColor: "#994cc3",
+    TagsColor: "#994cc3",
+    StringsColor: "#c96765",
+    OperatorsColor: "#994cc3",
+    AttributesColor: "#0c969b",
+    NumbersColor: "#aa0982",
+    ParametersColor: "#0c969b",
+
+    // Custom Color
+    SecondText: "#000000",
+    ContrastBackground: "#FFFFFF",
+};
+
+export const ColorThemes = {
+    MaterialThemeOceanic,
+    MaterialThemePalenight,
+    MaterialThemeDarker,
+    MaterialThemeDeepOcean,
+    ThemeOneLight,
+    ThemeMonoKaiPro,
+    MaterialThemeLighter,
+    SolarizedLight,
+    ThemeDracula,
+    ThemeArcDark,
+    ThemeOneDark,
+    ThemeNightOwl,
+    ThemeLightOwl,
+};
